requirements.txt
pyqt5==5.15.11
torch==2.2.2+cu118
numpy==1.23.5
tqdm==4.65.0
librosa==0.11.0
opencv-python==4.7.0.72
matplotlib==3.10.6
sympy==1.13.3
beartype==0.18.5
dill==0.3.7
protobuf==6.32.1
setuptools==78.1.1
scipy==1.15.3
torchvision==0.17.2+cu118
numba==0.60.0
urllib3==2.5.0
pillow==10.4.0
requests==2.32.5
six==1.17.0
soundfile==0.13.1
audioread==3.0.1
pydub==0.25.1
onnxruntime==1.22.1
moviepy==2.1.0
hdbscan==0.8.40
scikit-learn==1.7.2
modelscope==1.29.2
ffmpeg-python==0.2.0
future==1.0.0
pandas==1.5.3
tiktoken==0.9.0
elevenlabs==2.10.0
dataset==1.6.2
sqlalchemy==1.4.54
pypinyin==0.54.0
cryptography==45.0.7
google-genai==1.16